copypastel_rules_database/requirements.txt
mysql-connector-python==8.0.19
dnspython==1.16.0
